Standardize package definitions for kernel
Simply moved `kernel/default.nix` to `kernel/module.nix`; after, I moved
`kernel/package.nix` to `kernel/default.nix`. The idea here is that the
core packaged product is the kernel, which the module configuration
consumes while configuring NixOS.

As in last commit, this will not be the module code's final home.