Reorganize files module-first with a separate asahi-overlay
`<m1-support/nix/m1-support/default.nix>` contains a module as the entry
point for end-user OS integration because that's the point of this repo:
it targets NixOS. Separately, we use Asahi packages for this integration.

To that end, this changeset does two things:

1. Separate prior implementation into
   `nix/m1-support/{modules,asahi-overlay}` sandboxes.
2. Have `nix/m1-support/default.nix` "glue" these two together to add
   configurable m1 support to NixOS.

tl;dr no real approach changes

Everything in this commit is either a zero-contents-changed file move or
a reference update in `<m1-support/nix/m1-support/default.nix>` glue.